Split implementation.
