+++ include/methods/methods.h
@@ -1,7 +1,7 @@
 #ifndef LIBYCVID_INCLUDE_METHODS_H
 #define LIBYCVID_INCLUDE_METHODS_H
 
-yc_vid_status_t yc_vid_view_init(yc_env_world_t *world, yc_vid_view_t **result);
+yc_vid_status_t yc_vid_view_initialize(yc_env_world_t *world, yc_vid_view_t **result);
 yc_vid_status_t yc_vid_view_destroy(yc_vid_view_t *view, yc_vid_renderer_t *renderer);
 
 yc_vid_status_t yc_vid_view_port_set(
